Issue #2080823 by alexpott, swentel, Wim Leers: Create API to discover content or config entities' soft dependencies and use this to present a confirm form on module uninstall.

+++ core/modules/language/lib/Drupal/language/Entity/Language.php
@@ -96,7 +96,7 @@ class Language extends ConfigEntityBase implements LanguageInterface {
   public static function preDelete(EntityStorageControllerInterface $storage_controller, array $entities) {
     $default_language = \Drupal::service('language.default')->get();
     foreach ($entities as $entity) {
-      if ($entity->id() == $default_language->id) {
+      if ($entity->id() == $default_language->id && !$entity->isUninstalling()) {
         throw new DeleteDefaultLanguageException('Can not delete the default language');
       }
     }
